chore(trading): update trading thunks structure
