+++ suite-common/trading/src/index.ts
@@ -7,6 +7,5 @@ export * from './constants';
 export * from './reducers/tradingReducer';
 export * from './reducers/buyReducer';
 export * from './middlewares/tradingMiddleware';
-export * from './thunks/buy';
-export * from './thunks/tradingThunks';
+export * from './thunks';
 export * from './hooks/useTradingInfo';
