+++ suite-common/trading/src/thunks/buy/__tests__/handleRequestThunk.test.ts
@@ -4,7 +4,7 @@ import { CryptoId } from 'invity-api';
 import { configureMockStore, extraDependenciesMock } from '@suite-common/test-utils';
 import { getNetwork } from '@suite-common/wallet-config';
 
-import { buyThunks } from '../';
+import { buyThunks } from '../../';
 import { ALTERNATIVE_QUOTES } from '../../../__fixtures__/buyUtils';
 import { invityAPI } from '../../../invityAPI';
 import { initialState, prepareTradingReducer } from '../../../reducers/tradingReducer';
