+++ suite-common/trading/src/thunks/buy/__tests__/loadInfoThunk.test.ts
@@ -3,7 +3,7 @@ import { FiatCurrenciesProps } from 'invity-api';
 
 import { configureMockStore } from '@suite-common/test-utils';
 
-import { buyThunks } from '../';
+import { buyThunks } from '../../';
 import { invityAPI } from '../../../invityAPI';
 import { buyInitialState } from '../../../reducers/__fixtures__/buyTradingReducer';
 import { tradingBuyReducer } from '../../../reducers/buyReducer';
