+++ suite-common/trading/src/thunks/buy/__tests__/selectQuoteThunk.test.ts
@@ -9,7 +9,7 @@ import {
 
 import { configureMockStore, extraDependenciesMock } from '@suite-common/test-utils';
 
-import { buyThunks } from '../';
+import { buyThunks } from '../../';
 import { MIN_MAX_QUOTES_OK } from '../../../__fixtures__/buyUtils';
 import { invityAPI } from '../../../invityAPI';
 import { BuyInfo, TradingBuyState } from '../../../reducers/buyReducer';
