--- suite-common/trading/src/thunks/buy/index.ts
@@ -1,11 +1,0 @@
-import { confirmTradeThunk } from './confirmTradeThunk';
-import { handleRequestThunk } from './handleRequestThunk';
-import { loadInfoThunk } from './loadInfoThunk';
-import { selectQuoteThunk } from './selectQuoteThunk';
-
-export const buyThunks = {
-    loadInfoThunk,
-    handleRequestThunk,
-    selectQuoteThunk,
-    confirmTradeThunk,
-};
