+++ suite-common/trading/src/thunks/common/__tests__/verifyAddressThunk.test.ts
@@ -5,10 +5,10 @@ import { configureMockStore, extraDependenciesMock } from '@suite-common/test-ut
 import { confirmAddressOnDeviceThunk, selectSelectedDevice } from '@suite-common/wallet-core';
 import { Account, AddressDisplayOptions } from '@suite-common/wallet-types';
 
-import { accounts } from '../../reducers/__fixtures__/account';
-import { tradingBuyActions } from '../../reducers/buyReducer';
-import { initialState, prepareTradingReducer } from '../../reducers/tradingReducer';
-import { tradingThunks } from '../tradingThunks';
+import { tradingThunks } from '../../';
+import { accounts } from '../../../reducers/__fixtures__/account';
+import { tradingBuyActions } from '../../../reducers/buyReducer';
+import { initialState, prepareTradingReducer } from '../../../reducers/tradingReducer';
 
 const tradingReducer = prepareTradingReducer(extraDependenciesMock);
 const mockedSuiteReducer = createReducerWithExtraDeps(
